RPT_model/detr/models/vim/vim_requirements.txt
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
# causal-conv1d @ file:///home/zhulianghui/VisionProjects/mamba/lib/causal_conv1d-1.0.0%2Bcu118torch2.1cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=79a4bab633ebff031e615d5e8ba396b0dc0c046f4406980ee238fb86a9090038
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
# contourpy==1.2.0
contourpy
cycler==0.12.1
databricks-cli==0.18.0
datasets==2.15.0
dill==0.3.7
docker==6.1.3
einops==0.7.0
entrypoints==0.4
filelock==3.13.1
Flask==3.0.0
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.2
gunicorn==21.2.0
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
Mako==1.3.0
# mamba-ssm @ file:///home/zhulianghui/VisionProjects/mamba/lib/mamba_ssm-1.0.1%2Bcu118torch2.1cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=71ad1b1eafb05a6e8a41fd82e046fe85511d6378fa3a583e55215b6aa1d65ab9
Markdown==3.5.1
MarkupSafe==2.1.3
# matplotlib==3.8.2
matplotlib
mlflow==2.9.1
mmcv==1.3.8
mmsegmentation==0.14.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
# networkx==3.2.1
networkx
ninja==1.11.1.1
# numpy==1.26.2
numpy

# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==8.9.2.26
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.18.1
# nvidia-nvjitlink-cu12==12.3.101
# nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
# pandas==2.1.3
pandas
Pillow==10.1.0
platformdirs==4.1.0
prettytable==3.9.0
protobuf==4.25.1
pyarrow==14.0.1
pyarrow-hotfix==0.6
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-hostlist==1.23.0
pytz==2023.3.post1
PyYAML==6.0.1
querystring-parser==1.2.4
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.3.2
# scipy==1.11.4
scipy
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.23
sqlparse==0.4.4
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
timm==0.4.12
tokenizers==0.15.0
tomli==2.0.1
# torch==2.1.1+cu118
# torchvision==0.16.1+cu118
tqdm==4.66.1
transformers==4.35.2
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
websocket-client==1.7.0
Werkzeug==3.0.1
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
